Yeah, turns out there is yet another config struct called config.
Basically, original problem was since there is 2 or more different structs
with the same name and there was one which was basically a class, on exit
the destructor which was handled in a TLS callback in C++ CRT because it
wasnt manually defined was called and corrupted thread memory.

tldr: have distinct struct/class names people for heaven's sake.